changed type of hash variables

+++ src/hash.h
@@ -1,5 +1,5 @@
 #ifndef HASH_H
 #define HASH_H
-long hash_count(int spaces, int max_X, int max_O);
-long hash(int* board, int size);
+int hash_count(int spaces, int max_X, int max_O);
+int hash(int* board, int size);
 #endif
